add utilities and waste water analytics events
